pyproject.toml
[project]
name = "agents-prod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "curl-cffi>=0.14.0",
    "google-adk>=1.27.1",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langsmith[google-adk]>=0.7.18",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "python-telegram-bot[job-queue,webhooks]>=22.7",
    "requests>=2.32.5",
]
